feat: cacheManager를 만들어서 이미지를 캐싱하도록 구현
